[CXP-2145] Remove migrate tests and move helpers to remaining containers tests
