--- pkg/process/checks/enabled_checks_linux_test.go
@@ -1,40 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build linux
-
-package checks
-
-import (
-	"testing"
-
-	configmock "github.com/DataDog/datadog-agent/pkg/config/mock"
-)
-
-func TestProcessEventsCheckEnabled(t *testing.T) {
-	deps := createDeps(t)
-	t.Run("default", func(t *testing.T) {
-		cfg := configmock.New(t)
-
-		enabledChecks := getEnabledChecks(t, cfg, configmock.NewSystemProbe(t), deps.WMeta, deps.GpuSubscriber, deps.NpCollector, deps.Statsd)
-		assertNotContainsCheck(t, enabledChecks, ProcessEventsCheckName)
-	})
-
-	t.Run("enabled", func(t *testing.T) {
-		cfg := configmock.New(t)
-		cfg.SetWithoutSource("process_config.event_collection.enabled", true)
-
-		enabledChecks := getEnabledChecks(t, cfg, configmock.NewSystemProbe(t), deps.WMeta, deps.GpuSubscriber, deps.NpCollector, deps.Statsd)
-		assertContainsCheck(t, enabledChecks, ProcessEventsCheckName)
-	})
-
-	t.Run("disabled", func(t *testing.T) {
-		cfg := configmock.New(t)
-		cfg.SetWithoutSource("process_config.event_collection.enabled", false)
-
-		enabledChecks := getEnabledChecks(t, cfg, configmock.NewSystemProbe(t), deps.WMeta, deps.GpuSubscriber, deps.NpCollector, deps.Statsd)
-		assertNotContainsCheck(t, enabledChecks, ProcessEventsCheckName)
-	})
-}
